adding new config field
